Fix mermaid graphs on document
I use Typora for my Markdown editor, which features the integration of a few graphing generators like Mermaid.js. Github doesn't have this, so instead of the images generated, it just showed the code on how to generate it. I made a MASTER document which contains this code, and a viewable document, sort of like an exported version of the MASTER.